Extract tests of Maybe
